refactor: make common entrypoint for different app modes
