+++ package.json
@@ -8,7 +8,7 @@
     "test": "npx jest",
     "build": "webpack --mode=production",
     "start:prod": "npm run build && node dist/bundle.js",
-    "start:multi": "npx ts-node  src/claster.ts"
+    "start:multi": "npx ts-node  src/index.ts --mode=lb"
   },
   "repository": {
     "type": "git",
